backend/requirements.txt
# Vollständige Lokal-Installation (mit Spracherkennung).
# Für die Render-Cloud-Demo wird die schlankere Version aus
# /requirements.txt im Projekt-Root verwendet (ohne faster-whisper).

flask==3.1.0
gunicorn==23.0.0
requests>=2.31.0
fpdf2>=2.8.0
python-docx>=1.1.0
python-dotenv>=1.0.0
faster-whisper>=1.0.0
